+++ src/components/buttons/slack-button.tsx
@@ -1,10 +1,8 @@
-import React from "react";
-import { Button } from "../common";
-import slackIcon from "../../images/slack-icon.svg";
+import React from 'react';
+import { Button } from '../common';
+const slackIcon = require('../../images/slack-icon.svg');
 
-export interface SlackButtonProps {}
-
-const GithubButton: React.FC<SlackButtonProps> = () => {
+const GithubButton = () => {
   return (
     <Button>
       <a
@@ -14,9 +12,9 @@ const GithubButton: React.FC<SlackButtonProps> = () => {
       >
         <div
           style={{
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
+            display: 'flex',
+            alignItems: 'center',
+            justifyContent: 'center',
           }}
         >
           <div>Join us on </div>
@@ -24,7 +22,7 @@ const GithubButton: React.FC<SlackButtonProps> = () => {
           <img
             src={slackIcon}
             height="65px"
-            style={{ padding: "-50px", margin: "-80px 0px" }}
+            style={{ padding: '-50px', margin: '-80px 0px' }}
           />
         </div>
       </a>