+++ src/components/common/index.tsx
@@ -1,2 +1,2 @@
-export { default as Button } from "./button";
-export { default as Section } from "./section";
+export { default as Button } from './button';
+export { default as Section } from './section';