+++ src/components/sections/about-site.tsx
@@ -1,23 +1,29 @@
-import React from "react";
-import { Section } from "../common";
-import { link } from "fs";
-
-export interface AboutSiteProps {}
+import React from 'react';
+import { Section } from '../common';
+const gatsbyIcon = require('../../images/gatsby-icon.svg');
+const typescriptIcon = require('../../images/typescript-icon.svg');
 
 const links = {
-  gatsby: "https://www.gatsbyjs.org",
-  typescript: "https://www.typescriptlang.org",
-  netlify: "",
+  gatsby: 'https://www.gatsbyjs.org',
+  typescript: 'https://www.typescriptlang.org',
+  netlify: '',
 };
 
-const AboutSite: React.SFC<AboutSiteProps> = () => {
+const AboutSite = () => {
   return (
     <Section bg="#efefef">
       <h2>About This Site</h2>
       Built with&nbsp;
-      <a href={links.gatsby}>Gatsby</a>
+      <a href={links.gatsby}>
+        <img height="20px" src={gatsbyIcon} />
+        Gatsby
+      </a>
       &nbsp;&amp;&nbsp;
-      <a href={links.typescript}>TypeScript</a>
+      <a href={links.typescript}>
+        {' '}
+        <img height="20px" src={typescriptIcon} />
+        TypeScript
+      </a>
     </Section>
   );
 };