+++ src/lambda/fetchNextMeetupEvent.ts
@@ -1,11 +1,11 @@
-const axios = require("axios");
-const { format } = require("date-fns");
+import axios from 'axios';
+import {format} from 'date-fns';
 
 exports.handler = async (event, context, callback) => {
   // TODO- Convert to environment variable.
   const MEETUP_API_EVENTS_URL =
-    "http://api.meetup.com/long-island-javascript-group/events?status=upcoming&page=1";
-  const send = body => {
+    'http://api.meetup.com/long-island-javascript-group/events?status=upcoming&page=1';
+  const send = (body) => {
     callback(null, {
       statusCode: 200,
       // headers: {
@@ -22,10 +22,10 @@ exports.handler = async (event, context, callback) => {
       const result = await axios.get(MEETUP_API_EVENTS_URL);
       const { data } = result;
       const nextEvent = data[0];
-      const date = format(nextEvent.local_date, "MMMM DD, YYYY");
-      const time = `${format(nextEvent.time, "h:mm")} - ${format(
+      const date = format(nextEvent.local_date, 'MMMM DD, YYYY');
+      const time = `${format(nextEvent.time, 'h:mm')} - ${format(
         nextEvent.time + nextEvent.duration,
-        "h:mm A"
+        'h:mm A',
       )}`;
       const rsvps = `${nextEvent.yes_rsvp_count} ${nextEvent.group.who}`;
       const venue = nextEvent.venue.name;
@@ -46,7 +46,7 @@ exports.handler = async (event, context, callback) => {
     }
   };
 
-  if (event.httpMethod === "GET") {
+  if (event.httpMethod === 'GET') {
     getNextMeetupEvent();
   }
 };